inference_service/requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

transformers==4.41.2
pillow==10.3.0
numpy==1.26.4

redis==5.0.4
sqlalchemy==2.0.30
pymysql==1.1.0
python-dotenv==1.0.1